+++ src/features/smart-trading/components/index.ts
@@ -7,4 +7,5 @@ export { LiveActivityFeed, CompactActivityFeed } from "./LiveActivityFeed";
 export { TrackedWalletsPanel } from "./TrackedWalletsPanel";
 export { RecentTradesPanel } from "./RecentTradesPanel";
 export { TransactionsPanel } from "./TransactionsPanel";
-
+export { PositionProgressCard } from "./PositionProgressCard";
+export { ActivePositionsPanel } from "./ActivePositionsPanel";